pyproject.toml
[tool.poetry]
name = "adafruit_io_cli"
version = "0.0.2"
description = "CLI for the adafruit IO API and python client"
authors = ["Paul <pjurek3@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "adafruit_io_cli"}]

[tool.poetry.scripts]
ada = "adafruit_io_cli.main:app"

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.7.0"}
rich = "^12.6.0"
httpx = "^0.23.1"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-benchmark = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
